requirements.txt
streamlit
pandas
google-generativeai
pymupdf
Pillow
openpyxl
gspread
oauth2client
